common/requirements.txt
# FastAPI and Server
fastapi==0.109.2
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Redis and Caching
redis==5.0.1
aioredis==2.0.1
fastapi-limiter==0.1.6

# HTTP Client
httpx==0.26.0

# Utils
pytz==2024.1 